refactor(backup): :recycle: move backup path resolution to core
